refactor(agents): import PromptContextMode from models layer
- Update all agents imports to use models.prompt_meta
- Break execution → agents circular dependency
- Complete migration of PromptContextMode to models layer
