+++ src/vibe3/agents/plan_prompt.py
@@ -16,8 +16,8 @@ from loguru import logger
 
 from vibe3.config.settings import VibeConfig
 from vibe3.exceptions import VibeError
-from vibe3.execution.prompt_meta import PromptContextMode
 from vibe3.models.plan import PlanRequest
+from vibe3.models.prompt_meta import PromptContextMode
 from vibe3.prompts.context_builder import PromptContextBuilder, make_context_builder
 from vibe3.prompts.manifest import PromptManifest, PromptProvider
 from vibe3.resources.runtime_assets import resolve_runtime_asset
