+++ src/vibe3/agents/review_prompt.py
@@ -20,7 +20,7 @@ from loguru import logger
 from vibe3.analysis.snapshot_diff_section import build_snapshot_diff_section
 from vibe3.config.settings import VibeConfig
 from vibe3.exceptions import VibeError
-from vibe3.execution.prompt_meta import PromptContextMode
+from vibe3.models.prompt_meta import PromptContextMode
 from vibe3.models.review import ReviewRequest
 from vibe3.prompts.context_builder import PromptContextBuilder, make_context_builder
 from vibe3.prompts.manifest import PromptManifest, PromptProvider
